+++ changebyus/static/src/coffee/views/partials-project/ProjectMemberListItemView.coffee
@@ -62,13 +62,15 @@ define ["underscore", "backbone", "jquery", "template", "abstract-view"],
             ### EVENTS ---------------------------------------------###
             delete:(e)-> 
                 e.preventDefault()
-                $.ajax(
-                    type: "DELETE"
-                    url: "/api/user/#{@model.id}"
-                ).done (res_)=>
-                    if res_.success
-                        @model.collection.remove @model
-                        @$el.remove() 
+                confirmation = confirm("Are you sure you want to delete #{@model.get('display_name')}: #{@model.get('first_name')} #{@model.get('last_name')}?");
+                if confirmation
+                    $.ajax(
+                        type: "DELETE"
+                        url: "/api/user/#{@model.id}"
+                    ).done (res_)=>
+                        if res_.success
+                            @model.collection.remove @model
+                            @$el.remove() 
 
             unflag:(e)-> 
                 e.preventDefault()
